+++ clients/java/src/main/java/io/lakefs/clients/api/RefsApi.java
@@ -387,6 +387,7 @@ public class RefsApi {
      * @param amount how many items to return (optional, default to 100)
      * @param objects list of paths, each element is a path of a specific object (optional)
      * @param prefixes list of paths, each element is a path of a prefix (optional)
+     * @param limit limit the number of items in return to &#39;amount&#39;. Without further indication on actual number of items. (optional)
      * @param _callback Callback for upload/download progress
      * @return Call to execute
      * @throws ApiException If fail to serialize the request body object
@@ -399,7 +400,7 @@ public class RefsApi {
         <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
      </table>
      */
-    public okhttp3.Call logCommitsCall(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, final ApiCallback _callback) throws ApiException {
+    public okhttp3.Call logCommitsCall(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, Boolean limit, final ApiCallback _callback) throws ApiException {
         Object localVarPostBody = null;
 
         // create path and map variables
@@ -429,6 +430,10 @@ public class RefsApi {
             localVarCollectionQueryParams.addAll(localVarApiClient.parameterToPairs("multi", "prefixes", prefixes));
         }
 
+        if (limit != null) {
+            localVarQueryParams.addAll(localVarApiClient.parameterToPair("limit", limit));
+        }
+
         final String[] localVarAccepts = {
             "application/json"
         };
@@ -448,7 +453,7 @@ public class RefsApi {
     }
 
     @SuppressWarnings("rawtypes")
-    private okhttp3.Call logCommitsValidateBeforeCall(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, final ApiCallback _callback) throws ApiException {
+    private okhttp3.Call logCommitsValidateBeforeCall(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, Boolean limit, final ApiCallback _callback) throws ApiException {
         
         // verify the required parameter 'repository' is set
         if (repository == null) {
@@ -461,7 +466,7 @@ public class RefsApi {
         }
         
 
-        okhttp3.Call localVarCall = logCommitsCall(repository, ref, after, amount, objects, prefixes, _callback);
+        okhttp3.Call localVarCall = logCommitsCall(repository, ref, after, amount, objects, prefixes, limit, _callback);
         return localVarCall;
 
     }
@@ -475,6 +480,7 @@ public class RefsApi {
      * @param amount how many items to return (optional, default to 100)
      * @param objects list of paths, each element is a path of a specific object (optional)
      * @param prefixes list of paths, each element is a path of a prefix (optional)
+     * @param limit limit the number of items in return to &#39;amount&#39;. Without further indication on actual number of items. (optional)
      * @return CommitList
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      * @http.response.details
@@ -486,8 +492,8 @@ public class RefsApi {
         <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
      </table>
      */
-    public CommitList logCommits(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes) throws ApiException {
-        ApiResponse<CommitList> localVarResp = logCommitsWithHttpInfo(repository, ref, after, amount, objects, prefixes);
+    public CommitList logCommits(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, Boolean limit) throws ApiException {
+        ApiResponse<CommitList> localVarResp = logCommitsWithHttpInfo(repository, ref, after, amount, objects, prefixes, limit);
         return localVarResp.getData();
     }
 
@@ -500,6 +506,7 @@ public class RefsApi {
      * @param amount how many items to return (optional, default to 100)
      * @param objects list of paths, each element is a path of a specific object (optional)
      * @param prefixes list of paths, each element is a path of a prefix (optional)
+     * @param limit limit the number of items in return to &#39;amount&#39;. Without further indication on actual number of items. (optional)
      * @return ApiResponse&lt;CommitList&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      * @http.response.details
@@ -511,8 +518,8 @@ public class RefsApi {
         <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
      </table>
      */
-    public ApiResponse<CommitList> logCommitsWithHttpInfo(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes) throws ApiException {
-        okhttp3.Call localVarCall = logCommitsValidateBeforeCall(repository, ref, after, amount, objects, prefixes, null);
+    public ApiResponse<CommitList> logCommitsWithHttpInfo(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, Boolean limit) throws ApiException {
+        okhttp3.Call localVarCall = logCommitsValidateBeforeCall(repository, ref, after, amount, objects, prefixes, limit, null);
         Type localVarReturnType = new TypeToken<CommitList>(){}.getType();
         return localVarApiClient.execute(localVarCall, localVarReturnType);
     }
@@ -526,6 +533,7 @@ public class RefsApi {
      * @param amount how many items to return (optional, default to 100)
      * @param objects list of paths, each element is a path of a specific object (optional)
      * @param prefixes list of paths, each element is a path of a prefix (optional)
+     * @param limit limit the number of items in return to &#39;amount&#39;. Without further indication on actual number of items. (optional)
      * @param _callback The callback to be executed when the API call finishes
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
@@ -538,9 +546,9 @@ public class RefsApi {
         <tr><td> 0 </td><td> Internal Server Error </td><td>  -  </td></tr>
      </table>
      */
-    public okhttp3.Call logCommitsAsync(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, final ApiCallback<CommitList> _callback) throws ApiException {
+    public okhttp3.Call logCommitsAsync(String repository, String ref, String after, Integer amount, List<String> objects, List<String> prefixes, Boolean limit, final ApiCallback<CommitList> _callback) throws ApiException {
 
-        okhttp3.Call localVarCall = logCommitsValidateBeforeCall(repository, ref, after, amount, objects, prefixes, _callback);
+        okhttp3.Call localVarCall = logCommitsValidateBeforeCall(repository, ref, after, amount, objects, prefixes, limit, _callback);
         Type localVarReturnType = new TypeToken<CommitList>(){}.getType();
         localVarApiClient.executeAsync(localVarCall, localVarReturnType, _callback);
         return localVarCall;